feat: support where clause
