+++ packages/operational-transformation/src/index.ts
@@ -2,4 +2,4 @@ export * from './client';
 export { Server } from './server';
 export * from './identity';
 export * from './operation';
-export { Parser } from './parser';
+export * from './parser';
